added Weapon

+++ Properties/Services/CharacterService/CharacterService.cs
@@ -1,34 +1,42 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using anders3.Data;
 using anders3.Dtos.Character;
 using anders3.Models;
 using AutoMapper;
+using Microsoft.AspNetCore.Http;
 using Microsoft.EntityFrameworkCore;
 
 namespace anders3.Properties.Services.CharacterService
 {
     public class CharacterService : ICharacterService
     {
-        
+
+        private readonly IHttpContextAccessor _httpContextAccessor;
+
         private readonly IMapper _mapper;
         private readonly DataContext _context;
-        public CharacterService(IMapper mapper, DataContext context)
+        public CharacterService(IMapper mapper, DataContext context, IHttpContextAccessor httpContextAccessor)
         {
+            _httpContextAccessor = httpContextAccessor;
             _context = context;
             _mapper = mapper;
 
         }
+
+        private int GetUserId() => int.Parse(_httpContextAccessor.HttpContext.User.FindFirstValue(ClaimTypes.NameIdentifier));
         public async Task<ServiceResponse<List<GetCharacterDto>>> AddCharacter(AddCharacterDto newCharacter)
         {
             ServiceResponse<List<GetCharacterDto>> serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
             Character character = _mapper.Map<Character>(newCharacter);
-            
+            character.User = await _context.Users.FirstOrDefaultAsync(u => u.Id == GetUserId());
+
             await _context.Characters.AddAsync(character);
             await _context.SaveChangesAsync();
-            serviceResponse.Data = (_context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c))).ToList();
+            serviceResponse.Data = (_context.Characters.Where(c => c.User.Id== GetUserId()).Select(c => _mapper.Map<GetCharacterDto>(c))).ToList();
             return serviceResponse;
         }
 
@@ -37,12 +45,21 @@ namespace anders3.Properties.Services.CharacterService
             ServiceResponse<List<GetCharacterDto>> serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
             try
             {
-                Character character = await _context.Characters.FirstAsync(c => c.Id == id);
+                Character character = await _context.Characters
+                    .FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());
 
+                if(character != null)
+                {
                 _context.Characters.Remove(character);
                 await _context.SaveChangesAsync();
-
-                serviceResponse.Data = (_context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c))).ToList();
+                serviceResponse.Data = (_context.Characters.Where(c => c.User.Id == GetUserId())
+                    .Select(c => _mapper.Map<GetCharacterDto>(c))).ToList();
+                }
+                else
+                {
+                    serviceResponse.Success=false;
+                    serviceResponse.Message="Character was not found!";
+                }
             }
             catch (Exception ex)
             {
@@ -52,10 +69,10 @@ namespace anders3.Properties.Services.CharacterService
             return serviceResponse;
         }
 
-        public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacters(int userId)
+        public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacters()
         {
             ServiceResponse<List<GetCharacterDto>> serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
-            List<Character> dbCharacters = await _context.Characters.Where(c => c.User.Id == userId).ToListAsync();
+            List<Character> dbCharacters = await _context.Characters.Where(c => c.User.Id == GetUserId()).ToListAsync();
             serviceResponse.Data = (dbCharacters.Select(c => _mapper.Map<GetCharacterDto>(c))).ToList();
             return serviceResponse;
         }
@@ -63,7 +80,8 @@ namespace anders3.Properties.Services.CharacterService
         public async Task<ServiceResponse<GetCharacterDto>> GetCharacterById(int id)
         {
             ServiceResponse<GetCharacterDto> serviceResponse = new ServiceResponse<GetCharacterDto>();
-            Character dbCharacters = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id);
+            Character dbCharacters = await _context.Characters
+                .FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());
             serviceResponse.Data = _mapper.Map<GetCharacterDto>(dbCharacters);
             return serviceResponse;
         }
@@ -73,8 +91,9 @@ namespace anders3.Properties.Services.CharacterService
             ServiceResponse<GetCharacterDto> serviceResponse = new ServiceResponse<GetCharacterDto>();
             try
             {
-                Character character = await _context.Characters.FirstOrDefaultAsync(c => c.Id == updateCharacter.Id);
-
+                Character character = await _context.Characters.Include(c => c.User).FirstOrDefaultAsync(c => c.Id == updateCharacter.Id);
+                if (character.User.Id == GetUserId())
+                {
                 character.Name = updateCharacter.Name;
                 character.Class = updateCharacter.Class;
                 character.Defense = updateCharacter.Defense;
@@ -86,6 +105,11 @@ namespace anders3.Properties.Services.CharacterService
                 await _context.SaveChangesAsync();
 
                 serviceResponse.Data = _mapper.Map<GetCharacterDto>(character);
+                }
+                else{
+                    serviceResponse.Success=false;
+                    serviceResponse.Message="Character wasent found";
+                }
             }
             catch (Exception ex)
             {
